fix(content-manager): address MCP draft-schema review follow-ups
Round-two review follow-ups on the MCP draft-write schemas.

Component values on update are now a union that mirrors how the Document
Service dispatches: `updateOrCreateComponent` runs a partial UPDATE when the
value carries an `id`, and CREATEs a fresh row (stripping any `id`) when it does
not. Previously a partial update made every nested field optional, so an id-less
`{ seo: {} }` passed Zod and then replaced a valid row with one missing its
required fields — recursive entity validation downgrades `required` to
`notNull`, which an absent key satisfies, so nothing caught it. The patch branch
requires `id` with optional contents; the create branch enforces required
fields. `id` is coerced so the advertised schema stays a plain number rather
than a nested anyOf.

Repeatable-component and dynamic-zone minima now follow the same draft
projection as scalar `min`/`minLength`, so a D&P draft can clear a `min: 1` list
the way the admin panel allows. `max` stays unconditional.

Required relations and media get their own configuration-neutral hint: their
enforcement depends on `api.documents.strictRelations`, which is off by default,
so reusing the scalar "the server enforces it when the entry is saved" wording
was false for most configurations.

Also drops a private issue key from a doc comment, and adds a wholesale-replace
hint to repeatable-component and dynamic-zone arrays on update, since the
Document Service deletes any existing row absent from the incoming array and the
schema cannot express that constraint.

The `buildDataSchema` suite moves to its own file (the original was ~3.2k
lines), with the shared mock Strapi instance, model builders and request context
extracted to `test-fixtures.ts`. That module sits outside `__tests__/` because
the jest preset collects everything under that directory as a suite.
